declaration form component
