+++ drr/src/UI/src/app/drif/drif-eoi/drif-eoi-form.ts
@@ -7,7 +7,6 @@ import {
   propArray,
   propObject,
   required,
-  requiredTrue,
 } from '@rxweb/reactive-form-validators';
 import {
   ContactDetails,
@@ -20,7 +19,7 @@ import {
   ProjectType,
   ProponentType,
 } from '../../../model';
-import { AuthorizedRepresentativeForm } from '../../shared/drr-auth-rep/auth-rep-form';
+import { DeclarationForm } from '../../shared/drr-declaration/drr-declaration-form';
 
 export class FundingInformationItemForm implements FundingInformation {
   @prop()
@@ -302,27 +301,6 @@ export class OtherSupportingInformationForm {
   }
 }
 
-export class DeclarationForm {
-  @required()
-  @propObject(ContactDetailsForm)
-  authorizedRepresentative?: AuthorizedRepresentativeForm =
-    new AuthorizedRepresentativeForm({});
-
-  @prop()
-  @required()
-  @requiredTrue()
-  authorizedRepresentativeStatement?: boolean;
-
-  @prop()
-  @required()
-  @requiredTrue()
-  informationAccuracyStatement?: boolean;
-
-  constructor(values: DeclarationForm) {
-    Object.assign(this, values);
-  }
-}
-
 export class EOIApplicationForm {
   @propObject(ProponentInformationForm)
   proponentInformation?: ProponentInformationForm =
@@ -351,6 +329,4 @@ export class EOIApplicationForm {
 
   @propObject(DeclarationForm)
   declaration?: DeclarationForm = new DeclarationForm({});
-
-  // TODO: have a factory method to create a new instance of this form
 }
