+++ drr/src/UI/src/app/shared/drr-declaration/drr-declaration-form.ts
@@ -1,0 +1,28 @@
+import {
+  prop,
+  propObject,
+  required,
+  requiredTrue,
+} from '@rxweb/reactive-form-validators';
+import { AuthorizedRepresentativeForm } from '../drr-auth-rep/auth-rep-form';
+
+export class DeclarationForm {
+  @required()
+  @propObject(AuthorizedRepresentativeForm)
+  authorizedRepresentative?: AuthorizedRepresentativeForm =
+    new AuthorizedRepresentativeForm({});
+
+  @prop()
+  @required()
+  @requiredTrue()
+  authorizedRepresentativeStatement?: boolean;
+
+  @prop()
+  @required()
+  @requiredTrue()
+  informationAccuracyStatement?: boolean;
+
+  constructor(values: DeclarationForm) {
+    Object.assign(this, values);
+  }
+}
